+++ src/main/java/fll/Launcher.java
@@ -30,6 +30,7 @@ import java.net.Socket;
 import java.net.URL;
 import java.nio.file.Files;
 import java.nio.file.Path;
+import java.nio.file.Paths;
 import java.sql.Connection;
 import java.sql.SQLException;
 import java.util.Arrays;
@@ -68,6 +69,7 @@ import org.apache.commons.cli.CommandLineParser;
 import org.apache.commons.cli.DefaultParser;
 import org.apache.commons.cli.HelpFormatter;
 import org.apache.commons.cli.Options;
+import org.apache.commons.io.FileUtils;
 import org.apache.commons.lang3.SystemUtils;
 import org.checkerframework.checker.initialization.qual.UnknownInitialization;
 import org.checkerframework.checker.nullness.qual.Nullable;
@@ -945,7 +947,7 @@ public class Launcher extends JFrame {
   /**
    * @return the directory or null if not found
    */
-  private @Nullable Path getSponsorLogosDirectory(@UnknownInitialization(Launcher.class) Launcher this) {
+  private static @Nullable Path getSponsorLogosDirectory() {
     final Path classesPath = TomcatLauncher.getClassesPath();
     final Path webroot = TomcatLauncher.findWebappRoot(classesPath);
     if (null == webroot) {
@@ -964,7 +966,7 @@ public class Launcher extends JFrame {
   /**
    * @return the directory or null if not found
    */
-  private @Nullable Path getSlideshowDirectory(@UnknownInitialization(Launcher.class) Launcher this) {
+  private static @Nullable Path getSlideshowDirectory() {
     final Path classesPath = TomcatLauncher.getClassesPath();
     final Path webroot = TomcatLauncher.findWebappRoot(classesPath);
     if (null == webroot) {
@@ -1236,6 +1238,34 @@ public class Launcher extends JFrame {
     } catch (final IOException e) {
       throw new FLLRuntimeException("Error creating temporary file", e);
     }
+
+    final Path oldBaseDir = Paths.get(oldInstallationDirectory);
+    final Path oldWebDir = oldBaseDir.resolve("web");
+
+    // copy sponsor logos
+    final Path oldSponsorLogos = oldWebDir.resolve("sponsor_logos");
+    final Path newSponsorLogos = getSponsorLogosDirectory();
+    if (null == newSponsorLogos) {
+      throw new FLLRuntimeException("Unable to find current sponsor logos directory");
+    }
+    try {
+      FileUtils.copyDirectory(oldSponsorLogos.toFile(), newSponsorLogos.toFile());
+    } catch (final IOException e) {
+      throw new FLLRuntimeException("Error copying sponsor logos", e);
+    }
+
+    // copy slideshow
+    final Path oldSlideshow = oldWebDir.resolve("slideshow");
+    final Path newSlideshow = getSlideshowDirectory();
+    if (null == newSlideshow) {
+      throw new FLLRuntimeException("Unable to find current slideshow directory");
+    }
+    try {
+      FileUtils.copyDirectory(oldSlideshow.toFile(), newSlideshow.toFile());
+    } catch (final IOException e) {
+      throw new FLLRuntimeException("Error copying slideshow", e);
+    }
+
   }
 
 }